+++ R/to_numeric.R
@@ -85,25 +85,23 @@ to_numeric.data.frame <- function(x,
     verbose = verbose
   )
 
-  # drop numerics, when append is not FALSE
+  # when we append variables, we call ".process_append()", which will
+  # create the new variables and updates "select", so new variables are processed
   if (!isFALSE(append)) {
+    # drop numerics, when append is not FALSE
     select <- colnames(x[select])[!vapply(x[select], is.numeric, FUN.VALUE = logical(1L))]
+    # process arguments
+    args <- .process_append(
+      x,
+      select,
+      append,
+      append_suffix = "_n",
+      keep_factors = TRUE
+    )
+    # update processed arguments
+    x <- args$x
+    select <- args$select
   }
-
-  # process arguments
-  args <- .process_std_args(
-    x,
-    select,
-    exclude,
-    weights = NULL,
-    append,
-    append_suffix = "_n",
-    force = TRUE
-  )
-
-  # update processed arguments
-  x <- args$x
-  select <- args$select
 
   out <- sapply(
     x[select],