+++ tests/testthat/test-data_reverse.R
@@ -51,6 +51,30 @@ test_that("reverse works with data frame", {
   )
 })
 
+test_that("reverse works with data frame and append", {
+  test <- data.frame(
+    x = 1:5,
+    y = c(3, 8, 2, 5, 1)
+  )
+  expect_identical(
+    reverse(test, select = "x", append = TRUE),
+    data.frame(
+      x = 1:5,
+      y = c(3, 8, 2, 5, 1),
+      x_r = as.double(5:1)
+    )
+  )
+  expect_identical(
+    reverse(test, append = TRUE),
+    data.frame(
+      x = 1:5,
+      y = c(3, 8, 2, 5, 1),
+      x_r = as.double(5:1),
+      y_r = c(6, 1, 7, 4, 8)
+    )
+  )
+})
+
 test_that("reverse: arg 'select' works with formula", {
   test <- data.frame(
     x = 1:5,
@@ -354,6 +378,32 @@ test_that("reverse works with data frames (grouped data)", {
       id = rep(c("A", "B"), each = 3),
       value1 = c(10, 10, 3, 6, 2, 3),
       value2 = c(4, 6, 3, 10, 6, 5),
+      stringsAsFactors = FALSE
+    )
+  )
+})
+
+
+test_that("reverse works with grouped data frames and append", {
+  skip_if_not_installed("poorman")
+
+  test <- data.frame(
+    x = 1:6,
+    y = c(3, 8, 2, 5, 1, 4),
+    grp = rep(c("a", "b"), 3),
+    stringsAsFactors = FALSE
+  )
+  expect_identical(
+    test %>%
+      poorman::group_by(grp) %>%
+      reverse(append = TRUE) %>%
+      poorman::ungroup(),
+    data.frame(
+      x = 1:6,
+      y = c(3, 8, 2, 5, 1, 4),
+      grp = rep(c("a", "b"), 3),
+      x_r = as.double(c(5, 6, 3, 4, 1, 2)),
+      y_r = as.double(c(1, 4, 2, 7, 3, 8)),
       stringsAsFactors = FALSE
     )
   )